refactor(brain): address self-review findings from the extraction pass
- DreamQuarantinedEntry moves into dream-plan.ts, removing the
  type-only circular import between dream-plan.ts and dream.ts before
  anyone adds a value import to it.
- Document on WIKILINK_ALIAS_RE why the unicode flag the entity
  extractor used is safe for the link extractor that formerly ran
  without it.
